backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15.0,<3.16
requests>=2.31,<3
django-cors-headers>=4.3,<5
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
whitenoise>=6.6,<7
